requirements.txt
numpy<2
flask==3.1.3
waitress==3.0.2
transformers==4.44.2
huggingface_hub==0.24.6
safetensors==0.4.5
python-dotenv==1.0.1
huggingface_hub==0.24.6
datasets==2.19.2
pillow
matplotlib
torchvision